+++ CHANGES.rst
@@ -8,7 +8,8 @@ Changelog
   downstream packages can add extra fields to ``@scadenziario-day`` event
   results without duplicating the whole endpoint.
   [fedevancin]
-
+- Fix ``@scadenziario-day`` endpoint to return preview image properly.
+  [daniele]
 
 6.3.16 (2026-04-07)
 -------------------